api/requirements.txt
fastapi==0.112.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
sagemaker==2.229.0
replicate==0.32.0
uvicorn==0.30.6